refactor(common): organize lib into submodules
